fix: image fetching and pagination bugs

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/persistence/TeamDao.java
@@ -17,9 +17,9 @@ public interface TeamDao {
 
     List<Long> getActiveTournaments(Long teamId, Integer page);
 
-    Integer getActivePages(Long team_id);
+    Long getActivePages(Long team_id);
 
-    Integer getPastPages(Long team_id);
+    Long getPastPages(Long team_id);
 
     List<Team> getUserTeams(Long user_id);
